requirements.txt
django-moneyfield
django-money==0.15
django-currencies
django-aggregate-if==0.5
django-mathfilters==0.4.0
appdirs==1.4.0
packaging==16.8
pyparsing==2.1.10
amqp==2.2.2
Babel==2.8.0
backports-abc==0.5
billiard==3.5.0.3
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.29
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cursor-pagination==0.1.3
django-datetime-widget==0.9.3
django-debug-toolbar==1.8
django-extensions==1.8.1
django-rest-swagger==2.2.0
#django-timezone-field==3.0
djangorestframework==3.7.7
flower==0.9.2
future==0.16.0
futures==3.1.1
gunicorn==19.9.0
idna==2.5
#importlib-metadata==0.23
itypes==1.1.0
Jinja2>=2.10.1
jsonfield==2.0.2
kombu==4.1.0
libmagic==1.0.0
MarkupSafe==1.1.0
meld3==1.0.2
#more-itertools==7.2.0
netaddr==0.7.19
openapi-codec==1.3.2
psycopg2-binary==2.7.5
#psycopg2-binary==2.7.5
#python-crontab==2.3.6
#python-dateutil==2.8.0
python-magic==0.4.15
#python-magic-bin==0.4.14
python-memcached==1.59
pytz==2018.3
requests==2.22.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.2.3
supervisor==3.3.3
#supervisor==4.1.0
thehive4py==1.4.3
tornado==4.5.1
tzlocal
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.12.2
#zipp==0.6.0
#django-filter==2.2.0
#django-cors-headers==3.2.0
#djangorestframework-simplejwt==4.4.0
django-session-timeout==0.0.4
Pillow==6.2.1
